fix: arbitrary point ok

+++ labs/lab_2/app/src/main/java/explan/model/ModelingResult.java
@@ -1,11 +1,11 @@
 package explan.model;
 
 public class ModelingResult {
-    public float rhoTheoretical;
-    public float rhoPractical;
-    public float avgWaitTime;
+    public double rhoTheoretical;
+    public double rhoPractical;
+    public double avgWaitTime;
 
-    public ModelingResult(float rhoTheoritical, float rhoPractical, float avgWaitTime) {
+    public ModelingResult(double rhoTheoritical, double rhoPractical, double avgWaitTime) {
         this.rhoTheoretical = rhoTheoritical;
         this.rhoPractical = rhoPractical;
         this.avgWaitTime = avgWaitTime;
