+++ labs/lab_2/app/src/main/java/explan/model/experiment/SimulationResult.java
@@ -4,19 +4,19 @@ import java.util.Deque;
 import java.util.List;
 
 public class SimulationResult {
-    public float totalTime;
+    public double totalTime;
     public Deque<Request> requestsQueued;
     public List<Request> requestsProcessed;
 
     public Request requestProcessing;
-    public float workerIdleTime;
+    public double workerIdleTime;
 
     public SimulationResult(
-            float totalTime,
+            double totalTime,
             Deque<Request> queue,
             List<Request> processed,
             Request processing,
-            float workerIdle) {
+            double workerIdle) {
         this.totalTime = totalTime;
         this.requestsQueued = queue;
         this.requestsProcessed = processed;
@@ -24,13 +24,13 @@ public class SimulationResult {
         this.workerIdleTime = workerIdle;
     }
 
-    public float statRho() {
+    public double statRho() {
         // return avgWaitTime() / totalTime;
         return statLambda() / statMu();
     }
 
-    public float avgWaitTime() {
-        float avgTime = 0.0f;
+    public double avgWaitTime() {
+        double avgTime = 0.0;
 
         for (var request : requestsProcessed)
             avgTime += request.waitingTime();
@@ -44,12 +44,12 @@ public class SimulationResult {
         return avgTime / totalRequestsHandled();
     }
 
-    private float statLambda() {
-        return (float) totalRequestsCome() / totalTime;
+    private double statLambda() {
+        return (double) totalRequestsCome() / totalTime;
     }
 
-    private float statMu() {
-        return (float) totalRequestsHandled() / totalTime;
+    private double statMu() {
+        return (double) totalRequestsHandled() / totalTime;
     }
 
     private int totalRequestsCome() {
